+++ contracts/burner/src/lib.rs
@@ -1,5 +1,2 @@
 pub mod contract;
 pub mod msg;
-
-#[cfg(target_arch = "wasm32")]
-cosmwasm_std::create_entry_points_with_migration!(contract);